requirements.txt
numpy
pandas
matplotlib
scipy
scikit-image
opencv-python
openpyxl
tqdm
glob2
pillow
pyyaml